Update system/modules/core/pages/PageRegular.php
On line 364 switch the two if Blocks, that Mootool is first, and then Jquery.
The solution is to use the array_unshift() function, to get the scripts to the right place. After that jquery is first, than Mootool and then all the other scripts you can add with "$GLOBALS['TL_JAVASCRIPT'][]".