Fixed: cacheMode == browser always sends 'no-cache'
When `$GLOBALS['TL_CONFIG']['cacheMode'] = 'browser';` and
while sending Headers, a checked value was always `null`.
Thus the Headers for client caching were always prohibting
clients to cache any HTML document.

This fix replaces the variable that is checked and where
it is used for setting the correct Header.

This fix is a port of a fix in Contao 2.11.10 for Contao 3.x.