Simple hack to fix instantiating the Files-singleton with the wrong concrete implementation. Note that this indeed is a hack and doesn't fix the root-issue, which is an incorrect call-flow.
Details:
To handle safemode installation the installer takes some effort
to only instantiate the Files-Object after FTP settings are saved in
order to prevent the instantiation of the wrong conrete
File-Implementation.

This, however, doesn't work because InstallTool->loadLanguageFile()
is called by InstalTool's constructor, and it in turn again will
result in a call to Files::getInstance and an instantiation of
Files\Php. Because Files is a singleton this will prevent
the instantiation of the (correct) Files\Ftp-Implementation later on.

The fix is to simply relax the singleton constraint, by in effect
moving to one singleton per handler-type.