Pass current ContentModel to getContentElement() istead of ID
Instead of passing the ID of a content element to `getContentElement` the already fetched row should be applyed to the `ContentModel`.

I need to customize the content element row and want to call the `tl_content::addCteType`. At the moment I have to reimplement the whole method because of this limitation.

Besides you do not have to create yet another database request (I know Contao does not really care about database requests in the backend).