Add Hook 'getAdaptiveImage'
Is it possible to add this Hook? I have an extension which resize Images for mobile Devices.
For my function I can't use the 'getImage'-Hook because it's called to late.
In this function I would like to implement own complex queries of data and many options for mobile websites.
My purpose is that the function works for all images whatever equal modules or elements or further extensions. 
Otherwise it makes no sense :)