Use fgets instead of fread
´fread()´ ignores newlines which can cause problems with very small class files (e.g. Model classes).
The preg_match won't be executed if the filehandler has reached EOF, resulting in ignoring the class. That will lead to missing classes in the `ide_compat.php`.

By using fgets the file reading will also stop when reaching a newline. In that case, even small class files will be added correctly.