Security hardening: Set secure cookie flag when using SSL
Even if a website is already accessible via HTTPS most setups provide a
fallback for HTTP users to redirect to HTTPS.

In this scenario Contao can leak sensible *_USER_AUTH cookies: Imagine a
user is in the same network like an attacker. If the attacker manages to
trick the user to access the Contao website via HTTP the attacker can
capture the user's session.

This commit mitigates the problem by activating the secure flag for cookies
when using SSL.